services/file_service/requirements.txt
Flask>=2.0
python-dotenv
firebase-admin
